\name{bayou-package}
\alias{bayou-package}
\alias{bayou}
\docType{package}
\title{
  bayou
}
\description{
  A package for inferring adaptive evolution to phylogenetic comparative data using Bayesian reversible-jump
  estimation of multi-optima Ornstein-Uhlenbeck models.
}
\details{
  \tabular{ll}{
    Package: \tab bayou\cr
    Type: \tab Package\cr
    License: \tab GPL version 2 or greater?\cr
  }
  
}
\author{
  JC Uyeda and LJ Harmon
  
  Maintainer: Josef C Uyeda <josef.uyeda@gmail.com>
}
