\name{bayou2OUwie}
\alias{bayou2OUwie}
\title{Converts bayou data into OUwie format}
\usage{
  bayou2OUwie(pars, tree, dat)
}
\arguments{
  \item{pars}{A list with parameter values specifying
  \code{sb} = the branches with shifts, \code{loc} = the
  location on branches where a shift occurs and \code{t2} =
  the optima to which descendants of that shift inherit}

  \item{tree}{A phylogenetic tree}

  \item{dat}{A vector of tip states}
}
\description{
  \code{OUwie2bayou} calculates the probability density of
  a value
}

