\name{computelnr}
\alias{computelnr}
\title{Compute marginal likelihood}
\usage{
  computelnr(Kchains, ssfits, Bk, samp)
}
\description{
  \code{computelnr} computes the marginal likelihood of a
  set of chains estimated via stepping stone sampling and
  produced by the function \code{steppingstone}
}

