% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayou-moves.R
\name{.adjustTheta}
\alias{.adjustTheta}
\title{Adjust a Randomly Selected Theta Parameter}
\usage{
.adjustTheta(
  cache,
  pars,
  d,
  type = "slidingwindow",
  move = NULL,
  ct = NULL,
  prior = NULL
)
}
\arguments{
\item{cache}{Cache object containing phylogenetic tree and trait data.}

\item{pars}{List of current parameter values.}

\item{d}{Tuning parameter for the proposal step size.}

\item{type}{Proposal type: "slidingwindow" or "multiplier".}

\item{move}{The parameter being modified.}

\item{ct}{Optional control object.}

\item{prior}{Optional prior function.}
}
\value{
A list containing new parameter values, Hastings ratio, and selected theta index.
}
\description{
This function modifies a randomly selected theta parameter using either a sliding window or multiplier proposal.
}
