% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayou-moves.R
\name{.proposalFn}
\alias{.proposalFn}
\title{General Proposal Function Caller}
\usage{
.proposalFn(u, ct, D, moves, cache, pars, prior = NULL)
}
\arguments{
\item{u}{A random value used to select the move.}

\item{ct}{Control object specifying the move probabilities.}

\item{D}{Tuning parameters for the proposals.}

\item{moves}{List of available proposal functions.}

\item{cache}{Cache object containing phylogenetic tree and trait data.}

\item{pars}{List of current parameter values.}

\item{prior}{Optional prior function for parameter values.}
}
\value{
A list containing new parameter values and Hastings ratio.
}
\description{
This function selects a move based on a random value `u` and executes the corresponding proposal function.
}
