% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayou-moves.R
\name{.slidespace}
\alias{.slidespace}
\title{Move a Shift in Space}
\usage{
.slidespace(j, pars, cache, ct, map)
}
\arguments{
\item{j}{Index of the shift being moved}

\item{pars}{The current parameter set}

\item{cache}{The cache list containing the tree and data}

\item{ct}{Control list governing constraints}

\item{map}{A map structure of the tree}
}
\value{
A list containing new shift branch positions and proposal probabilities
}
\description{
Moves a shift up or down within or between branches in the phylogeny.
}
