% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayou-moves.R
\name{.slidingWindowProposal}
\alias{.slidingWindowProposal}
\title{Sliding Window Proposal Function}
\usage{
.slidingWindowProposal(cache, pars, d, move, ct = NULL, prior = NULL)
}
\arguments{
\item{cache}{Cache object containing phylogenetic tree and trait data.}

\item{pars}{List of current parameter values.}

\item{d}{Tuning parameter for the proposal step size.}

\item{move}{The parameter being modified.}

\item{ct}{Optional control object.}

\item{prior}{Optional prior function.}
}
\value{
A list containing new parameter values and Hastings ratio.
}
\description{
This function generates a sliding window proposal for a given parameter.
}
