% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayou-moves.R
\name{.splitmerge}
\alias{.splitmerge}
\title{Perform a Split-Merge Move on the Phylogenetic Tree}
\usage{
.splitmerge(pars, cache, d, ct, move = NULL, prior = NULL)
}
\arguments{
\item{pars}{The current parameter set}

\item{cache}{The cache list containing the tree and data}

\item{d}{The tuning parameter}

\item{ct}{Control list governing constraints}

\item{move}{The move being performed (unused)}

\item{prior}{The prior function (unused)}
}
\value{
A list with updated parameters and the Hastings ratio
}
\description{
Adds or removes a shift along a branch of the phylogenetic tree.
}
