% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayou-moves.R
\name{.vectorMultiplier}
\alias{.vectorMultiplier}
\title{Vector Sliding Window Proposal}
\usage{
.vectorMultiplier(cache, pars, d, move, ct = NULL, prior = NULL)
}
\arguments{
\item{cache}{Cached data.}

\item{pars}{Current parameters.}

\item{d}{Tuning parameter.}

\item{move}{Name of the parameter being adjusted.}

\item{ct}{Control list.}

\item{prior}{Optional prior function.}
}
\value{
A list containing the new parameters and Hastings ratio.
}
\description{
Adjusts one element in a vector parameter using a sliding window.
}
