\name{edgemap.W}
\alias{.edgemap.W}
\alias{edgemap.W}
\title{Calculate the weight matrix of a set of regimes on a phylogeny using an edge map.}
\usage{
  edgemap.W(tree, pars, emap, alpha = NULL)

  .edgemap.W(cache, pars, emap, TotExp, alpha = NULL)
}
\description{
  Calculate the weight matrix of a set of regimes on a
  phylogeny using an edge map.
}

