\name{identify.branches}
\alias{identify.branches}
\title{Identify shifts on branches of a phylogenetic tree}
\usage{
  identify.branches(tree, n, fixed.loc = TRUE,
    plot.simmap = TRUE)
}
\description{
  \code{identify.branches} opens an interactive phylogeny
  plot that allows the user to specify the location of
  shifts in a phylogenetic tree.
}

