% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayou-likelihood.R
\docType{data}
\name{model.OU}
\alias{model.OU}
\title{Bayou Models}
\format{
An object of class \code{list} of length 9.
}
\usage{
model.OU
}
\description{
Default bayou models. New models may be specified by providing a set of moves, control weights,
tuning parameters, parameter names, RJ parameters and a likelihood function.
}
\keyword{datasets}
