% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/conversion-utilities.R
\name{OU.repar}
\alias{OU.repar}
\title{Calculates the alpha and sigma^2 from a parameter list with supplied phylogenetic half-life and stationary variance}
\usage{
OU.repar(pars)
}
\arguments{
\item{pars}{A bayou formatted parameter list with parameters halflife (phylogenetic halflife) and Vy (stationary variance)}
}
\value{
A list with values for alpha and sig2.
}
\description{
Calculates the alpha and sigma^2 from a parameter list with supplied phylogenetic half-life and stationary variance
}

