\name{OUwie2bayou}
\alias{OUwie2bayou}
\title{Converts OUwie data into bayou format}
\usage{
  OUwie2bayou(tree, trait)
}
\arguments{
  \item{tree}{A phylogenetic tree with states at internal
  nodes as node labels}

  \item{trait}{A data frame in OUwie format}
}
\description{
  \code{OUwie2bayou} calculates the probability density of
  a value
}

