% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bayou-mcmc-utilities.R
\name{overparameterize.startingPoint}
\alias{overparameterize.startingPoint}
\title{Generate an overparameterized starting point for the MCMC}
\usage{
overparameterize.startingPoint(prior, tree, dat)
}
\arguments{
\item{prior}{A prior function}

\item{tree}{A phylogenetic tree of class 'phylo'}

\item{dat}{A named data vector}
}
\description{
This function takes a prior function and generates a starting point that can be entered for \code{startpar}
in the function \code{bayou.mcmc}
}
\details{
This function creates an "overparameterized" starting point for running the mcmc. It gives n-1 tips a unique
optimum close to the actual data value. This is useful if you expect steep likelihood peaks that may be hard to find,
as these often will be easier to access from this overparameterized model. Generally, the overparameterization will have
a very high likelihood and a very low prior.
}

