\name{phenogram.density}
\alias{phenogram.density}
\title{Plot a pheongram with the posterior density for optima values}
\usage{
  phenogram.density(tree, dat, burnin = 0, chain,
    colors = NULL, pp.cutoff = NULL, K = NULL, ...)
}
\description{
  Plot a pheongram with the posterior density for optima
  values
}

