\name{plot.bayouMCMC}
\alias{plot.bayouMCMC}
\title{S3 method for plotting bayouMCMC objects}
\usage{
  \method{plot}{bayouMCMC} (x, ...)
}
\arguments{
  \item{x}{A mcmc chain of class 'bayouMCMC' produced by
  the function bayou.mcmc and loaded into the environment
  using load.bayou}

  \item{...}{Additional arguments passed to
  \code{plot.mcmc} from the \code{coda} package}
}
\description{
  S3 method for plotting bayouMCMC objects
}

