% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teaching-functions.R
\name{plotOUtreesim}
\alias{plotOUtreesim}
\title{A function to visualize a multi-optimum OU process evolving on a phylogeny}
\usage{
plotOUtreesim(pars, tree, ptsperunit = 100, pal = rainbow, aph = 255, lwd = 1)
}
\arguments{
\item{pars}{A bayou parameter list to simulate the OU process from}

\item{tree}{A phylogenetic tree}

\item{ptsperunit}{A number giving the number of points to simulate per unit time}

\item{pal}{A color palette function}

\item{aph}{The alpha value for transparency of the lines}

\item{lwd}{The width of the lines}
}
\value{
**No return value**, called for **side effects**.
The function **generates a plot** showing the simulated trait evolution over time.
}
\description{
A function to visualize a multi-optimum OU process evolving on a phylogeny
}
