% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bayou-plotting.R
\name{plotRegimes}
\alias{plotRegimes}
\title{Function to plot the regimes from a simmap tree}
\usage{
plotRegimes(tree, col = NULL, lwd = 1, pal = rainbow, ...)
}
\arguments{
\item{tree}{A simmap tree of class phylo or simmap with a tree$maps list}

\item{col}{A named vector of colors to assign to character states, if NULL, then colors are generated from pal}

\item{lwd}{A numeric value indicating the width of the edges}

\item{pal}{A color palette function to generate colors if col=NULL}

\item{...}{Optional arguments that are passed to plot.phylo}
}
\description{
Function to plot the regimes from a simmap tree
}
\details{
This function uses plot.phylo to generate coordinates and plot the tree, but plots the
'maps' element of phytools' simmap format. This provides much of the functionality of plot.phylo from
the ape package. Currently, only types 'phylogram', 'unrooted', 'radial', and 'cladogram' are allowed. Phylogenies must
have branch lengths.
}

