% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayou-mcmc-utilities.R
\name{print.bayouFit}
\alias{print.bayouFit}
\title{S3 method for printing bayouFit objects}
\usage{
\method{print}{bayouFit}(x, ...)
}
\arguments{
\item{x}{A 'bayouFit' object produced by \code{bayou.mcmc}}

\item{...}{Additional parameters passed to \code{print}}
}
\value{
**No return value**, called for **side effects**.
This function prints a summary of a **bayou model fit**, including
posterior probabilities, model parameters, and key statistics.
}
\description{
S3 method for printing bayouFit objects
}
