% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayou-mcmc-utilities.R
\name{print.bayouMCMC}
\alias{print.bayouMCMC}
\title{S3 method for printing bayouMCMC objects}
\usage{
\method{print}{bayouMCMC}(x, ...)
}
\arguments{
\item{x}{A mcmc chain of class 'bayouMCMC' produced by the function bayou.mcmc and loaded into the environment using load.bayou}

\item{...}{Additional arguments}
}
\description{
S3 method for printing bayouMCMC objects
}
