% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayou-steppingstone.R
\name{print.ssMCMC}
\alias{print.ssMCMC}
\title{S3 method for printing ssMCMC objects}
\usage{
\method{print}{ssMCMC}(x, ...)
}
\arguments{
\item{x}{An ssMCMC object}

\item{...}{Optional arguments passed to print}
}
\value{
**No return value**, called for **side effects**.
This function prints a summary of an `ssMCMC` object, including the estimated marginal likelihood,
the number of power posteriors run, and the log marginal likelihood contributions across steps.
}
\description{
S3 method for printing ssMCMC objects
}
