% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayou-mcmc-utilities.R
\name{pull.pars}
\alias{pull.pars}
\title{Utility function for retrieving parameters from an MCMC chain}
\usage{
pull.pars(i, chain, model = "OU")
}
\arguments{
\item{i}{An integer giving the sample to retrieve}

\item{chain}{A bayouMCMC chain}

\item{model}{The parameterization used, either "OU", "QG" or "OUrepar"}
}
\value{
A bayou formatted parameter list
}
\description{
Utility function for retrieving parameters from an MCMC chain
}
