\name{smOU.lik}
\alias{OU.lik}
\alias{smOU.lik}
\title{Calculate the likelihood of a multi-optima OU model from simmap format regimes}
\usage{
  smOU.lik(pars, tree, X, SE = 0, model = "OU")

  OU.lik(pars, tree, X, SE = 0, model = "OU")
}
\description{
  Calculates the likelihood of an OU model with regimes
  stored in tree$maps.

  Calculates the likelihood of an OU model with regimes
  specified by a parameter list
}

