% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayou-deprecated.R
\name{bayou-deprecated}
\alias{bayou-deprecated}
\title{Deprecated functions in package bayou.}
\value{
This documentation does not return a value. It serves as a reference for deprecated functions in the \code{bayou} package.
}
\description{
The functions listed below are deprecated and will be defunct in
  the near future. When possible, alternative functions with similar
  functionality are also mentioned. Help pages for deprecated functions are
  available at \code{help("-deprecated")}.
}
\keyword{internal}
