% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayou-package.R
\name{bayou-package}
\alias{bayou-package}
\alias{bayou}
\title{Bayesian Fitting of Ornstein-Uhlenbeck Models to Phylogenies}
\description{
A package for inferring adaptive evolution to phylogenetic 
comparative data using Bayesian reversible-jump estimation of 
multi-optima Ornstein-Uhlenbeck models.
}
\details{
bayou-package
}
\author{
Josef C Uyeda
}
