% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bayou-package.R
\name{bayou-package}
\alias{bayou}
\alias{bayou-package}
\title{bayou-package}
\description{
A package for inferring adaptive evolution to phylogenetic
comparative data using Bayesian reversible-jump
estimation of multi-optima Ornstein-Uhlenbeck models.
}
\author{
Josef C Uyeda
}

