% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/conversion-utilities.R
\name{bayou2OUwie}
\alias{bayou2OUwie}
\title{Converts bayou data into OUwie format}
\usage{
bayou2OUwie(pars, tree, dat)
}
\arguments{
\item{pars}{A list with parameter values specifying \code{sb} = the branches with shifts,
\code{loc} = the location on branches where a shift occurs and \code{t2} = the optima to which
descendants of that shift inherit}

\item{tree}{A phylogenetic tree}

\item{dat}{A vector of tip states}
}
\value{
A list with an OUwie formatted tree with mapped regimes and an OUwie formatted data table
}
\description{
\code{bayou2OUwie} Converts a bayou formatted parameter list into OUwie formatted tree and data table that can be analyzed in OUwie
}

