% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evangeline.R
\name{Evangeline}
\alias{Evangeline}
\title{The Theme Evangeline}
\usage{
Evangeline(gridline_x = FALSE, gridline_y = TRUE)
}
\arguments{
\item{gridline_x}{If you want grid lines on your x axis in your visualization, set gridline_x=TRUE, if you do not want them enter gridline_x=FALSE}

\item{gridline_y}{If you want grid lines on your y axis on your visualization, set gridline_y=TRUE, if you do not want them enter gridline_y=FALSE}
}
\value{
A visualization with a centered title and subtitle will show up. An additional feature of this theme is the x axis titles being titled and centered.
}
\description{
The Theme Evangeline
}
\examples{
Theme_Evangeline has examples on how the theme is used
}
