% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayou-plotting.R
\name{makeTransparent}
\alias{makeTransparent}
\title{Make a color transparent (Taken from an answer on StackOverflow by Nick Sabbe)}
\usage{
makeTransparent(someColor, alpha = 100)
}
\arguments{
\item{someColor}{A color, either a number, string or hexidecimal code}

\item{alpha}{The alpha transparency. The maxColorValue is set to 255.}
}
\value{
A character vector of colors in hexadecimal format with the specified transparency applied.
}
\description{
Make a color transparent (Taken from an answer on StackOverflow by Nick Sabbe)
}
