% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayou-plotting.R
\name{plot.bayouMCMC}
\alias{plot.bayouMCMC}
\title{S3 method for plotting bayouMCMC objects}
\usage{
\method{plot}{bayouMCMC}(x, ...)
}
\arguments{
\item{x}{A mcmc chain of class 'bayouMCMC' produced by the function bayou.mcmc and loaded into the environment using load.bayou}

\item{...}{Additional arguments passed to \code{plot.mcmc} from the \code{coda} package}
}
\value{
No return value, called for side effects. This function generates diagnostic
trace and density plots for MCMC chains of class `bayouMCMC` to assess convergence
and parameter distributions.
}
\description{
S3 method for plotting bayouMCMC objects
}
