% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayou-plotting.R
\name{plotBayoupars}
\alias{plotBayoupars}
\title{Plot parameter list as a simmap tree}
\usage{
plotBayoupars(pars, tree, ...)
}
\arguments{
\item{pars}{A bayou formatted parameter list}

\item{tree}{A tree of class 'phylo'}

\item{...}{Additional arguments passed to plotRegimes}
}
\value{
No return value, called for side effects. This function generates
a visualization of a phylogenetic tree with mapped regimes or other parameters.
}
\description{
Plot parameter list as a simmap tree
}
