\name{plotBayoupars}
\alias{plotBayoupars}
\title{Plot parameter list as a simmap tree}
\usage{
  plotBayoupars(pars, tree, ...)
}
\arguments{
  \item{pars}{A bayou formatted parameter list}

  \item{tree}{A tree of class 'phylo'}

  \item{...}{Additional arguments passed to plotSimmap}
}
\description{
  Plot parameter list as a simmap tree
}

