\name{print.bayouFit}
\alias{print.bayouFit}
\title{S3 method for printing bayouFit objects}
\usage{
  \method{print}{bayouFit} (x, ...)
}
\arguments{
  \item{x}{A 'bayouFit' object produced by
  \code{bayou.mcmc}}

  \item{...}{Additional parameters passed to \code{print}}
}
\description{
  S3 method for printing bayouFit objects
}

