% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayou-utilities.R
\name{print.priorFn}
\alias{print.priorFn}
\title{S3 method for printing priorFn objects}
\usage{
\method{print}{priorFn}(x, ...)
}
\arguments{
\item{x}{A function of class 'priorFn' produced by \code{make.prior}}

\item{...}{Additional arguments passed to \code{print}}
}
\value{
**No return value**, called for **side effects**.
This function prints a summary of a **priorFn** object, including
the expected model, required parameter list, and prior distribution functions.
}
\description{
S3 method for printing priorFn objects
}
