% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayou-utilities.R
\name{print.refFn}
\alias{print.refFn}
\title{S3 method for printing refFn objects}
\usage{
\method{print}{refFn}(x, ...)
}
\arguments{
\item{x}{A function of class 'refFn' produced by make.refFn}

\item{...}{Additional arguments passed to \code{print}}
}
\value{
**No return value**, called for **side effects**.
This function prints a summary of a **refFn** object, including
the expected model, required parameter list, and reference distribution functions.
}
\description{
S3 method for printing refFn objects
}
