% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bayou-steppingstone.R
\name{print.ssMCMC}
\alias{print.ssMCMC}
\title{S3 method for printing ssMCMC objects}
\usage{
\method{print}{ssMCMC}(x, ...)
}
\arguments{
\item{x}{An ssMCMC object}

\item{...}{Optional arguments passed to print}
}
\description{
S3 method for printing ssMCMC objects
}

