% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayou-simulation.r
\name{priorSim}
\alias{priorSim}
\title{Simulates parameters from bayou models}
\usage{
priorSim(prior, tree, plot = TRUE, nsim = 1, shiftpars = "theta", ...)
}
\arguments{
\item{prior}{A prior function created by \code{bayou::make.prior}}

\item{tree}{A tree of class 'phylo'}

\item{plot}{A logical indicating whether the simulated parameters should be plotted}

\item{nsim}{The number of parameter sets to be simulated}

\item{shiftpars}{A vector of parameters that split upon a shift, default is "theta"}

\item{...}{Parameters passed on to \code{plotSimmap(...)}}
}
\value{
A list of bayou parameter lists drawn from the prior.
}
\description{
\code{priorSim} Simulates parameters from the prior distribution specified by \code{make.prior}
}
