% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trumpets.R
\name{Trumpets}
\alias{Trumpets}
\title{Title}
\usage{
Trumpets(data, column_name)
}
\arguments{
\item{data}{This is the title of the data frame which you want to work in}

\item{column_name}{This is the name of column that you want to create a separate column for its order identification}
}
\description{
Title
}
