% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trumpets.R
\name{Trumpets}
\alias{Trumpets}
\title{Trumpets}
\usage{
Trumpets(data, column_name)
}
\arguments{
\item{data}{This is the title of the data frame which you want to work in}

\item{column_name}{This is the name of column that you want to create a separate column for its order identification}
}
\value{
This returns a data frame that more rearranges the objects based on the desired column as well as includes the new column with the numbers 1-number of rows
}
\description{
Trumpets
}
\examples{
BayouDataMoves has  an example of this function in action
}
